Refactor. - Created and used ConfigStrings resource. - Moved all the enums to Enums class and deleted the other classes. - Created ExceptionMessage class for the Forms.
