+++ Vueling/Vueling.Common.Logic/Enums/Enums.cs
@@ -6,4 +6,11 @@
         English = 1,
         Spanish = 2
     }
+    public enum Extension
+    {
+        TXT = 1,
+        JSON,
+        XML,
+        SQL
+    }
 }
