--- Vueling/Vueling.Common.Logic/Enums/ExtensionesFicheros.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Vueling.Common.Logic.Enums
-{
-    public static class ExtensionesFicheros
-    {
-        public enum Extension
-        {
-            TXT = 1, JSON, XML, SQL
-        }
-    }
-}
