+++ Vueling/Vueling.Common.Logic/Properties/ConfigStrings.resx
@@ -123,7 +123,22 @@
   <data name="English" xml:space="preserve">
     <value>en-GB</value>
   </data>
+  <data name="FileType" xml:space="preserve">
+    <value>tipoFichero</value>
+  </data>
+  <data name="JsonFile" xml:space="preserve">
+    <value>ListadoAlumnos.json</value>
+  </data>
+  <data name="Language" xml:space="preserve">
+    <value>Language</value>
+  </data>
   <data name="Spanish" xml:space="preserve">
     <value>es-ES</value>
+  </data>
+  <data name="TxtFile" xml:space="preserve">
+    <value>ListadoAlumnos.txt</value>
+  </data>
+  <data name="XmlFile" xml:space="preserve">
+    <value>ListadoAlumnos.xml</value>
   </data>
 </root>
